backend/requirements.txt
fastapi
uvicorn
python-dotenv
supabase
langgraph
langchain
langchain-core
langchain-google-genai
pydantic
httpx
